// src/notes/dto/create-note.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsBoolean, IsEnum, IsNotEmpty, IsOptional, IsString } from 'class-validator';

export class CreateNoteDto {
  @ApiProperty({
    description: 'Note timestamp',
    example: '2023-12-03T10:30:00',
  })
  @IsString()
  @IsNotEmpty()
  time: string;

  @ApiProperty({
    description: 'Note content',
    example: 'This is an important note',
  })
  @IsString()
  @IsNotEmpty()
  content: string;

  @ApiProperty({
    description: 'Note importance level',
    enum: ['medium', 'critical'],
    example: 'medium',
  })
  @IsEnum(['medium', 'critical'])
  @IsNotEmpty()
  importance: 'medium' | 'critical';

  @ApiProperty({
    description: 'Indicates if the note was generated by AI',
    example: false,
    required: false,
  })
  @IsBoolean()
  @IsOptional()
  isAiGenerated?: boolean;
}